% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{linal_embedded_active_db}
\alias{linal_embedded_active_db}
\title{The active database's name}
\usage{
linal_embedded_active_db(db)
}
\arguments{
\item{db}{A `linal_embedded_db` from `linal_embedded_db()`.}
}
\value{
A `character` scalar.
}
\description{
The active database's name
}
