% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linaldb.embedded-package.R
\docType{package}
\name{linaldb.embedded-package}
\alias{linaldb.embedded}
\alias{linaldb.embedded-package}
\title{linaldb.embedded: Embedded/Native R Binding for LINALDB}
\description{
Wraps the LINALDB engine's Rust `TensorDb` directly via `extendr` --
no HTTP server, no `linal serve` process. See `../EMBEDDED_CONTRACT.md`
for the Rust/R result-shape contract this package implements against,
and the `linaldb` package under `clients/r` for the HTTP-client
alternative (talks to a running `linal serve` instance instead).
}
\author{
\strong{Maintainer}: gorigami \email{develop@gorigami.xyz}

Authors:
\itemize{
  \item gorigami \email{develop@gorigami.xyz}
}

}
\keyword{internal}
